Alternation
Changes to accommodate time cell algorithms for sessions where the mouse
turned both left and right. This commit is functional for at least
01_28_2016, but unknown unknowns exist.

Known issues:
Need to standardize axes in plotTimeCells.
Significance markers appearing for inactive neurons.
Lots of time fields towards the end of the run, possible issue with this
session.